Modify session handling a bit
It stores a single cookie token for a session and authenticates the user by checking it at the `sessions` table and retrieving the UID from it.
If the token is stolen, the session would be stolen aswell. Is this okay..?
I don't know much about sessions yet. This will need a re-write, I suppose.